requirements.txt
Flask==2.2.5
requests==2.31.0
geopy==2.3.0
waitress==2.1.2
gunicorn==21.2.0
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
Flask-Bcrypt==1.0.1
email-validator==2.0.0
Werkzeug==2.2.3
